fix: compile warnings and use generated cmake file (#1)

+++ src/lib/vector2.c
@@ -4,7 +4,6 @@
  *--------------------------------------------------------------------------------------------*/
 #include <basal/math.h>
 #include <basal/vector2.h>
-#include <tiny-libc/tiny_libc.h>
 
 BlVector2 blVector2Make(float a, float b)
 {
@@ -81,7 +80,7 @@ float blVector2Dot(BlVector2 a, BlVector2 b)
 
 BlVector2 blVector2Reflect(BlVector2 incoming, BlVector2 normal)
 {
-    int reflectDot = blVector2Dot(incoming, normal);
+    float reflectDot = blVector2Dot(incoming, normal);
     return blVector2Sub(incoming, blVector2Scale(normal, reflectDot * 2));
 }
